Add loadOrCreate helper and learn to do basic variable assignment
